+++ src/features/base/components/mainLayout/mainLayout.tsx
@@ -41,7 +41,13 @@ import {DefaultCountry, EnvPattern} from "@lib/constants";
 import smartlookClient from "smartlook-client";
 import {setProgress} from "@features/progressUI";
 import {setUserId, setUserProperties} from "@firebase/analytics";
-import {isAppleDevise, isSupported, useInvalidateQueries, useMedicalEntitySuffix} from "@lib/hooks";
+import {
+    isAppleDevise,
+    isSupported,
+    useCalculateCnxSpeed,
+    useInvalidateQueries,
+    useMedicalEntitySuffix
+} from "@lib/hooks";
 import {fetchAndActivate, getRemoteConfig, getString} from "firebase/remote-config";
 import {useRequestQueryMutation} from "@lib/axios";
 import useMutateOnGoing from "@lib/hooks/useMutateOnGoing";
@@ -71,7 +77,7 @@ function MainLayout({...props}) {
     const {urlMedicalEntitySuffix} = useMedicalEntitySuffix();
     const {trigger: mutateOnGoing} = useMutateOnGoing();
     const {trigger: invalidateQueries} = useInvalidateQueries();
-    //useCalculateCnxSpeed(); // Check speed connection
+    useCalculateCnxSpeed(); // Check speed connection
     const audio = useMemo(() => new Audio("/static/sound/beep.mp3"), []);
 
     const {appointmentTypes} = useAppSelector(dashLayoutSelector);
@@ -310,22 +316,22 @@ function MainLayout({...props}) {
 
     useEffect(() => {
         const remoteConfig = getRemoteConfig(firebaseCloudSdk.firebase);
-        if (typeof window !== "undefined" && window?.Upscope && general_information && process.env.NODE_ENV !== 'development') {
+        if (typeof window !== "undefined" && window?.UpscopeLoad && general_information && process.env.NODE_ENV !== 'development') {
             fetchAndActivate(remoteConfig).then(() => {
                 const config = JSON.parse(getString(remoteConfig, 'medlink_remote_config'));
                 if (config.upscope) {
+                    window.UpscopeLoad(window, window.Upscope, document);
                     window.Upscope('init');
                     window.Upscope('updateConnection', {
                         // Set the  user ID below. If you don't have one, set to undefined.
                         uniqueId: general_information.uuid,
-
                         // Set the  username or email below (e.g. ["John Smith", "john.smith@acme.com"]).
                         identities: [roles[0], `${general_information.firstName} ${general_information.lastName}`]
                     });
                 }
             });
         }
-    }, [window?.Upscope, general_information]); // eslint-disable-line react-hooks/exhaustive-deps
+    }, [window?.UpscopeLoad, general_information]); // eslint-disable-line react-hooks/exhaustive-deps
 
     useEffect(() => {
         // Update notifications popup