+++ src/lib/hooks/useCalculateCnxSpeed.ts
@@ -1,46 +1,30 @@
-import {useEffect, useRef} from "react";
-import {useSnackbar} from "notistack";
-import {useTranslation} from "next-i18next";
+import {useCallback, useEffect} from "react";
+import {useAppDispatch} from "@lib/redux/hooks";
+import {setConnexionStatus} from "@features/base";
 
 function useCalculateCnxSpeed() {
-    const {enqueueSnackbar} = useSnackbar();
+    const dispatch = useAppDispatch();
 
-    const {t} = useTranslation("common");
+    const checkConnectionSpeed = useCallback(() => {
+        const imageAddr = `/static/img/absence.png`;
+        let startTime: number, endTime;
+        const downloadSize = 4096;
+        const download = new Image();
+        startTime = (new Date()).getTime();
+        download.src = imageAddr;
+        download.onload = function () {
+            endTime = (new Date()).getTime();
+            const duration = (endTime - startTime) / 1000;
+            const bitsLoaded = downloadSize * 8;
+            const speedBps = parseFloat((bitsLoaded / duration).toFixed(2));
+            const speedKbps = parseFloat((speedBps / 1024).toFixed(2));
 
-    const startTime = useRef<any>(null);
-    const endTime = useRef<any>(null);
-    const imageSize = useRef<any>(null);
-    const image = useRef<any>(new Image());
-    const speedInMbps = useRef<any>(null);
-    let imageLink = "https://source.unsplash.com/random?topics=nature";
-
-    const calculateSpeed = () => {
-        let timeDuration = (endTime.current - startTime.current) / 1000;
-        let loadedBits = imageSize.current * 8;
-        let speedInBps = parseFloat((loadedBits / timeDuration).toFixed(2));
-        let speedInKbps = parseFloat((speedInBps / 1024).toFixed(2));
-        speedInMbps.current = parseFloat((speedInKbps / 1024).toFixed(2));
-        if (speedInMbps.current < 0.6) {
-            enqueueSnackbar(t("slow-connection"), {
-                key: "slow",
-                variant: 'warning',
-                anchorOrigin: {horizontal: "left", vertical: "bottom"},
-                persist: true
-            })
-        }
-    }
-
-    image.current.onload = async function () {
-        endTime.current = new Date().getTime();
-        await fetch(imageLink).then((response) => {
-            imageSize.current = response.headers.get("content-length");
-            calculateSpeed();
-        });
-    };
+            dispatch(setConnexionStatus(speedKbps < 8))
+        };
+    }, [dispatch]);
 
     useEffect(() => {
-        startTime.current = new Date().getTime();
-        image.current.src = imageLink;
+        checkConnectionSpeed();
     }, []); // eslint-disable-line react-hooks/exhaustive-deps
 }